v3.0.0 feat. Astro! 🚀
Everything is pixel perfect as always... (save ✈️) but there is this one nasty bug where the entire app NEEDS to be in a wrapping div within body, and it needs to own the styles, otherwise it's a pixel or two off from the prod version. Nothing too serious, but I've decided to stick with the wrapping div for now. Nothing, and absolutely nothing I tried worked. Just wrapping the div and sticking the styles onto body/:root doesn't work too, they HAVE to be on the wrapping div. The styles themselves shouldn't affect the children in any way at all, and the user agent stylesheet is sufficiently reset. Well... it's a mystery for another day :P

Found another bug in Astro, reported it here:
https://github.com/withastro/astro/issues/3184